Moved file system management and random string code to appropriate files.
File system code is now in fs.{h|c}. This includes:

    - spool directory setup
    - metadata file management
    - pcap file creation and writing

Random hex string generation is now in str.h.
